implementação de Repository,Service e Controller para a entidade Cliente .

+++ api/src/main/java/com/loja/api/model/Cliente.java
@@ -21,7 +21,7 @@ public class Cliente {
 
     private String nome;
     private String email;
-    private int telefone;
-    private int cpf;
+    private Long telefone;
+    private Long cpf;
 
 }
